rename & move unsupported Windows C# tests, with updated descriptions
(and updated test snapshots)
