+++ tests/unsupported-platforms/windows-csharp/with-options/windows-csharp-with-options.test.js
@@ -1,4 +1,4 @@
-const lib = require('../../../../../../lib/lib.js');
+const lib = require('../../../../lib/lib.js');
 
 // special compact mocks for this test
 // with mock uuid for Windows:
@@ -21,10 +21,15 @@ jest.mock('fs-extra', () => ({
   },
 }));
 
-test('create alice-bobbi module using mocked lib with defaults on Windows', () => {
+test('unsupported Windows C# template with options', () => {
   const options = {
     platforms: ['windows'],
     name: 'alice-bobbi',
+    githubAccount: 'alicebits',
+    authorName: 'Alice',
+    authorEmail: 'contact@alice.me',
+    license: 'ISC',
+    namespace: 'Carol',
   };
 
   return lib(options)
